# backend/openapi.yaml
openapi: 3.0.3
info:
  title: HR / E-Learning API
  version: "1.0.0"
  description: |
    OpenAPI spec generated from your NestJS controllers & DTOs (User, Auth, Admin, Feedback, AuditLog, Backup).
    Sensitive fields (passwordHash, mfaSecret, mfaBackupCodes) are intentionally excluded from public response schemas.
servers:
  - url: http://localhost:3786
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuditEvent:
      type: string
      enum:
        - EMAIL_VERIFIED
        - USER_REGISTERED
        - USER_CHANGED_EMAIL
        - OTP_SENT
        - OTP_SEND_FAILED
        - LOGIN_SUCCESS
        - LOGIN_FAILED
        - LOGOUT
        - PASSWORD_RESET_REQUESTED
        - PASSWORD_RESET_COMPLETED
        - MFA_ENABLED
        - MFA_DISABLED
        - PASSWORD_CHANGED
        - PROFILE_UPDATED
        - UNAUTHORIZED_ACCESS
        - TOKEN_BLACKLISTED
        - RBAC_DENIED
        - ADMIN_CREATED_USER
        - ADMIN_USER_LOCK
        - ADMIN_USER_UNLOCK
        - ADMIN_FORCE_LOGOUT
        - ADMIN_ANNOUNCE_ALL
        - ROLE_CHANGED
        - ADMIN_ANNOUNCE_ROLE
        - ADMIN_DELETE_USER
        - DATA_EXPORT
    CreateAuditLogDto:
      type: object
      properties:
        userId:
          type: string
        event:
          $ref: '#/components/schemas/AuditEvent'
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
      required:
        - event
    UpdateAuditLogDto:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/AuditEvent'
        details:
          type: object
          additionalProperties: true
    PublicAuditDto:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        event:
          $ref: '#/components/schemas/AuditEvent'
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time

    # -------------------------
    # Admin DTOs
    # -------------------------
...
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId
paths:
...
  /audit:
    post:
      tags:
        - audit
      summary: Create an audit log entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuditLogDto'
      responses:
        '201':
          description: Created audit log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAuditDto'
    get:
      tags:
        - audit
      summary: List audit entries (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: string
        - name: event
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicAuditDto'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
                  limit:
                    type: integer
  /audit/{id}:
    get:
      tags:
        - audit
      summary: Get a single audit entry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Audit entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAuditDto'
    patch:
      tags:
        - audit
      summary: Update an audit entry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuditLogDto'
      responses:
        '200':
          description: Updated audit entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAuditDto'
    delete:
      tags:
        - audit
      summary: Delete an audit entry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /audit/purge/older-than/{days}:
    delete:
      tags:
        - audit
      summary: Purge audit entries older than X days
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Purge result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedCount:
                    type: integer
  /audit/security/failed-logins:
    get:
      tags:
        - audit
      summary: List failed login events
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Failed logins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuditDto'
  /audit/security/unauthorized:
    get:
      tags:
        - audit
      summary: List unauthorized access events
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Unauthorized access events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuditDto'
  /audit/event/{event}:
    get:
      tags:
        - audit
      summary: List audit events by type
      security:
        - bearerAuth: []
      parameters:
        - name: event
          in: path
          required: true
          schema:
            type: string
            enum:
              - EMAIL_VERIFIED
              - USER_REGISTERED
              - USER_CHANGED_EMAIL
              - OTP_SENT
              - OTP_SEND_FAILED
              - LOGIN_SUCCESS
              - LOGIN_FAILED
              - LOGOUT
              - PASSWORD_RESET_REQUESTED
              - PASSWORD_RESET_COMPLETED
              - MFA_ENABLED
              - MFA_DISABLED
              - PASSWORD_CHANGED
              - PROFILE_UPDATED
              - UNAUTHORIZED_ACCESS
              - TOKEN_BLACKLISTED
              - RBAC_DENIED
              - ADMIN_CREATED_USER
              - ADMIN_USER_LOCK
              - ADMIN_USER_UNLOCK
              - ADMIN_FORCE_LOGOUT
              - ADMIN_ANNOUNCE_ALL
              - ROLE_CHANGED
              - ADMIN_ANNOUNCE_ROLE
              - ADMIN_DELETE_USER
              - DATA_EXPORT
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Audit logs by event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuditDto'
  /audit/user/{userId}:
    get:
      tags:
        - audit
      summary: List audit events for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Audit logs for user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuditDto'

  # -------------------------
  # Admin
  # -------------------------
...